Reorganize settings file
