+++ tilavarauspalvelu/utils/get_version.py
@@ -1,0 +1,26 @@
+# ruff: noqa
+
+import subprocess  # nosec
+
+
+def get_git_revision_hash() -> str:
+    """
+    Retrieve the git hash for the underlying git repository or die trying
+    We need a way to retrieve git revision hash for sentry reports
+    I assume that if we have a git repository available we will
+    have git-the-comamand as well
+    """
+    try:
+        # We are not interested in gits complaints
+        git_hash = subprocess.check_output(  # nosec
+            ["git", "rev-parse", "HEAD"], stderr=subprocess.DEVNULL, encoding="utf8"
+        )
+    # i.e. "git" was not found
+    # should we return a more generic meta hash here?
+    # like "undefined"?
+    except FileNotFoundError:
+        git_hash = "git_not_available"
+    except subprocess.CalledProcessError:
+        # Ditto
+        git_hash = "no_repository"
+    return git_hash.rstrip()
